Time: 71 ms (71.11%), Space: 108.6 MB (83.6%) - LeetHub

+++ 2586-longest-square-streak-in-an-array/2586-longest-square-streak-in-an-array.cpp
@@ -6,7 +6,7 @@ public:
         int ans=0;
         for(auto i:s){          
             int tmp=1;  //tmp streak
-            long long x=i; //current
+            int x=i; //current
             while(s.find(x)!=s.end()){
                 int root=sqrt(x);
                 if(root*root==x && s.find(root)!=s.end()){//perfect square
